+++ src/xnatuploader/put.py
@@ -63,8 +63,6 @@ def resource(session, scan, *filenames, **kwargs):
         The ID of the subject to upload the dataset to
     scan_id : str
         The ID for the scan (defaults to the scan type)
-    date : str
-        The date of the scan
     modality : str
         The modality of the session to upload
     user : str
@@ -96,9 +94,6 @@ def resource(session, scan, *filenames, **kwargs):
     resource_name = kwargs.pop("resource_name", None)
     project_id = kwargs.pop("project_id", None)
     subject_id = kwargs.pop("subject_id", None)
-    date = kwargs.pop("date", None)
-    scanner_manufacturer = kwargs.pop("manufacturer", None)
-    scanner_model = kwargs.pop("model", None)
     scan_id = kwargs.pop("scan_id", None)
     modality = kwargs.pop("modality", None)
     if sanitize_re.match(session):
@@ -158,12 +153,6 @@ def resource(session, scan, *filenames, **kwargs):
                 # exist
                 xsubject = login.classes.SubjectData(label=subject_id, parent=xproject)
                 xsession = session_cls(label=session, parent=xsubject)
-                if date is not None:
-                    xsession.date = date
-                if scanner_manufacturer is not None:
-                    xsession.scanner.manufacturer = scanner_manufacturer
-                if scanner_model is not None:
-                    xsession.scanner.model = scanner_model
             else:
                 raise XnatUtilsNoMatchingSessionsException(
                     "'{}' session does not exist, to automatically create it "